Update translations for charging state connect_cable
Fixes #137

We may have misunderstood what Skoda and their API mean with a charging
status of `CONNECT_CABLE`

We've been assuming it means "You need to connect the cable" but instead
it probably means "There is a cable connected but that's it, we're
neither charging nor ready to charge".

A better translation is "No Power".

Note that the binary_sensor `charger_connected` already tells us exactly
if a charge cable is connected or not.

Translations based on google translate ...